% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate.R
\name{calc_tscore}
\alias{calc_tscore}
\title{Calculate a GAS T-score}
\usage{
calc_tscore(scores, weights = NULL, rho = 0.3, na.rm = TRUE)
}
\arguments{
\item{scores}{Discrete goal scores.}

\item{weights}{Optional goal weights. If not provided, all weights are
set to 1 (i.e. unweighted goals).}

\item{rho}{Average degree of inter-correlation between goal scores.
By default, \code{rho} = 0.3.}

\item{na.rm}{Logical. If TRUE (default), then removes missing values.}
}
\value{
A single score aggregating a subject's set of a goal scores.
}
\description{
Computes a single GAS T-score for a subject at a point in time.
}
\examples{
d <- list(scores = c(-2, 0, -1, 2, 0, 1),
          weights = c(3, 5, 1, 4, 2, 6))
calc_tscore(d$scores, d$weights)
}
