% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/thresholds.R
\name{create_thresholds}
\alias{create_thresholds}
\title{Create score thresholds}
\usage{
create_thresholds(n_levels = 5, centre = 0)
}
\arguments{
\item{n_levels}{The number of levels to use. Defaults to the traditional
5 levels.}

\item{centre}{The centre of the normal distribution from which the thresholds
are taken. See 'Details'.}
}
\value{
A numeric vector of thresholds with \code{n_levels + 1} values. The first
value will be \code{-Inf}, and the last \code{Inf}, so that extremely small or large
values can still be discretized.
}
\description{
Creates a set of thresholds which can be used to discretize
latent continuous goal scores into "observed" discrete scores.
}
\details{
This function follows the approach introduced by
Urach et al. 2019 to get equally spaced attainment levels from
a cumulative standard normal distribution.
The distribution \code{centre} can be shifted to adjust the "difficulty" of
goals.
}
\examples{
create_thresholds()
create_thresholds(n_levels = 3)

# Make goals slightly easier by shifting the thresholds left
create_thresholds(centre = -0.2)
# Or harder by shifting right
create_thresholds(centre = 0.2)

}
\references{
\insertRef{Urach2019}{gasr}
}
