% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/thresholds.R
\name{discretize_from_thresholds}
\alias{discretize_from_thresholds}
\title{Discretize from thresholds}
\usage{
discretize_from_thresholds(score_continuous, thresholds)
}
\arguments{
\item{score_continuous}{Numeric vector of continuous goal scores.}

\item{thresholds}{Numeric vector of thresholds separating the goal
attainment levels. Must be an even number of thresholds (which corresponds
to an odd number of attainment levels)}
}
\value{
Numeric vector, the same length as \code{score_continuous}, of integers
corresponding to discrete attainment levels. For a typical 5-point
attainment scale, the returned values will be of the set
\{-2, -1, 0, 1, 2\}.
}
\description{
Discretizes continuous latent goal score values into discrete levels
centred at 0, using a set of thresholds.
}
\examples{
thresh <- create_thresholds()
y <- rnorm(10)
y_discrete <- discretize_from_thresholds(y, thresh)
}
