% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{generate_subject_ids}
\alias{generate_subject_ids}
\title{Generate subject IDs}
\usage{
generate_subject_ids(n_subjects, prefix = "S")
}
\arguments{
\item{n_subjects}{The number of subjects.}

\item{prefix}{The prefix of the IDs. "S" by default.}
}
\value{
A character vector of length \code{n_subjects}.
}
\description{
Provides sequential numbered subject IDs for a specified number of subjects.
}
\examples{
generate_subject_ids(100)
generate_subject_ids(20, "subj-")

}
