% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate.R
\name{sim_goals}
\alias{sim_goals}
\title{Simulate goals}
\usage{
sim_goals(
  subjects,
  n_goals = NULL,
  n_goals_range = NULL,
  n_goals_prob = NULL,
  sigma_e = 0.5
)
}
\arguments{
\item{subjects}{A dataframe with a row per subject.}

\item{n_goals}{The exact number of goals for each subject.}

\item{n_goals_range}{A vector defining the minimum and maximum number of
goals per subjects. Randomly samples the range of goals with uniform
probbility.}

\item{n_goals_prob}{A named list with element \code{n_goals} defining the possible
numbers of goals, and element \code{prob} defining the probabilities of getting
each number. For example:
\verb{n_goals_prob = list(n_goals = 1:3, probs = c(0.3, 0.3, 0.4}.}

\item{sigma_e}{Standard deviation of the goal-level random effect, i.e.
the random error of each latent goal score.}
}
\value{
The same \code{subjects} dataframe with a \code{n_goals} column, and
\code{goals} list column. The \code{goals} list column contains tibbles with two
variables: a subject-specific goal index \code{goal_num}, and a normally
distributed random effect \code{goal_re}.
}
\description{
Simulate goals
}
\examples{
sim_subjects() \%>\%
  sim_goals(n_goals = 3, sigma_e = 0.8)

sim_subjects() \%>\%
  sim_goals(n_goals_range = c(2, 7))

sim_subjects() \%>\%
  sim_goals(n_goals_prob = list(n_goals = c(6, 3, 4),
                                prob = c(0.3, 0.5, 0.2)))

}
