% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate.R
\name{sim_subjects}
\alias{sim_subjects}
\title{Simulate subjects}
\usage{
sim_subjects(
  n_subjects = 100,
  sigma_u = 0.5,
  group_allocation = list(control = 0.5, treatment = 0.5)
)
}
\arguments{
\item{n_subjects}{Number of subjects.}

\item{sigma_u}{Standard deviation of the subject-level random effect.}

\item{group_allocation}{A named list of groups and their probability of
assignment. By default, will equally allocate subjects to "control" and
"treatment"}
}
\value{
A dataframe with a row for each subject.
}
\description{
Simulate subjects
}
\examples{
sim_subjects(n_subjects = 20)
sim_subjects(
  group_allocation = list("placebo" = 1/2,
                          "1x dose" = 1/4, "2x dose" = 1/4)
)
}
