% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate.R
\name{sim_treatment_effect}
\alias{sim_treatment_effect}
\title{Simulate a treatment effect}
\usage{
sim_treatment_effect(goals, delta = 0.3)
}
\arguments{
\item{goals}{A dataframe with a row for each goal.}

\item{delta}{The size of the treatment effect}
}
\value{
The same \code{goals} dataframe with the added \code{treatment_fe} column.
}
\description{
Currently only simulates a uniform distribution of treatment effects,
regardless of group allocation
}
