% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate.R
\name{sim_treatment_effect}
\alias{sim_treatment_effect}
\title{Simulate a treatment effect}
\usage{
sim_treatment_effect(goals, delta = 0.3)
}
\arguments{
\item{goals}{A dataframe with a row for each goal.}

\item{delta}{The size of the treatment effect}
}
\value{
The same \code{goals} dataframe with the added \code{treatment_fe} column.
}
\description{
Currently only simulates a uniform distribution of treatment effects,
regardless of group allocation.
}
\examples{
library(dplyr)
library(tidyr)
library(purrr)

# Generate some subjects and goals
d <- sim_subjects(n_subjects = 40) \%>\%
  sim_goals(n_goals_range = c(1, 3))

# Apply treatment effects in one of two ways
# Using purrr::map()
d \%>\%
  mutate(
    goals = map(
      goals,
      ~sim_treatment_effect(., delta = 0.4)
    )
  )

# Or using tidyr::unnest() on the goals column
d \%>\%
  unnest(goals) \%>\%
  group_by(subject_id) \%>\%
  sim_treatment_effect(delta = 0.4)

}
