% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate.R
\name{sim_trial}
\alias{sim_trial}
\title{Simulate a single cross-sectional trial}
\usage{
sim_trial(
  n_subjects = 40,
  sigma_u = 0.5,
  group_allocation = list(control = 0.5, treatment = 0.5),
  n_goals = NULL,
  n_goals_range = c(3, 6),
  n_goals_prob = NULL,
  sigma_e = 0.5,
  delta = 0.3,
  weight_type = "unweighted",
  n_levels = 5,
  centre = 0
)
}
\arguments{
\item{n_subjects}{Number of subjects.}

\item{sigma_u}{Standard deviation of the subject-level random effect.}

\item{group_allocation}{A named list of groups and their probability of
assignment. By default, will equally allocate subjects to "control" and
"treatment"}

\item{n_goals}{The exact number of goals for each subject.}

\item{n_goals_range}{A vector defining the minimum and maximum number of
goals per subjects. Randomly samples the range of goals with uniform
probability.}

\item{n_goals_prob}{A named list with element \code{n_goals} defining the possible
numbers of goals, and element \code{prob} defining the probabilities of getting
each number. For example:
\verb{n_goals_prob = list(n_goals = 1:3, probs = c(0.3, 0.3, 0.4)}.}

\item{sigma_e}{Standard deviation of the goal-level random effect, i.e.
the random error of each latent goal score.}

\item{delta}{The size of the treatment effect.}

\item{weight_type}{The type of weights to apply. See 'Details'.}

\item{n_levels}{The number of levels to use. Defaults to the traditional
5 levels.}

\item{centre}{The centre of the normal distribution from which the thresholds
are taken. See 'Details'.}
}
\value{
A data frame with a row per trial (numbered 1 to \code{n_sim}) which c
list column\code{ data} containing the simulated data.
}
\description{
Simulate a single cross-sectional trial
}
